+++ src/components/buttons/index.ts
@@ -1,0 +1,4 @@
+import Button from "./button";
+import ButtonBase from "./button-base";
+
+export { Button, ButtonBase };